[virtual_total] use has_attribute? for size aggregate
problem: attribute_present? was false for a nil valued attribute
fix: has_attribute is true for a nil valued attribute

As for size, select count(*) will always return a number,
  so the attribute can't be nil
and in ruby [].size will always return a number.
  so || 0 is probably not necessary nor is the try
But left them there incase it is a delegate or some special case due to
  a bug in development

TODO:  Test to confirm this?  Otherwise, is it relevant to the PR?

Sidenote:  Message notes are pulled from the original commit... but WTF is up
with the formatting on that second paragraph!?