[virtual_total] Fix nil for loaded rel.blank?
Note:  Untrue statement

    problem: fell back to lookup clause for nil value
    fix: falls back to default clause instead (0)

    - kbrock

In reality, turns out that what fixes this spec inconsistency is that
since we aren't ever checking for a `.blank?` on the relation, after the
`[].map.compact` call, we will get back a `[]`, so `.size == 0` as a
result.

Since `[]` will be the result of `rel.map {}` when `rel.blank?`, this
means the `rel.blank?` was not doing anything of use, and could be
removed.

TODO:  I assume that `rel` will always be a `[]`, correct?  Otherwise
instead of removing the `rel.blank?`, you could just change the default
to `0`.